% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_stream_num_trades}
\alias{backtest_stream_num_trades}
\title{The number of closed trades in a streaming backtest so far.}
\usage{
backtest_stream_num_trades(bt)
}
\arguments{
\item{bt}{A \code{wickra_stream} handle.}
}
\value{
A length-one numeric vector.
}
\description{
The number of closed trades in a streaming backtest so far.
}
