% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_version}
\alias{backtest_version}
\title{The native wickra-backtest library version.}
\usage{
backtest_version()
}
\value{
A length-one character vector.
}
\description{
The native wickra-backtest library version.
}
